vlm/api/requirements.txt
transformers==4.38.0
accelerate
tiktoken
einops
transformers_stream_generator==0.0.4
scipy
torchvision
pillow
tensorboard
matplotlib
flask
PyJWT